Added new main
New main was added to accept the cli interface.
